18 LLM Powered Chatbot/requirements.txt
flask
google-cloud-aiplatform
python-dotenv